port SetupList to ts
we need db types merged before we proceed
